Improving the External Competition Algorithm
The External Competition protocol was defined to not support recipes with values more than 1. so I solved the issue and now it supports.
while testing, i figured out that python makes fractional sum of 0 to a smallest number other than 0. also fixed this bug by updating the number of sum to be 0 in case it is in under 0.00000000001.
I also created additional python files to find out the input that makes the bug of optimal number lower than SBB algorithm.